pyproject.toml
[project]
name = "avero"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain>=1.2.15",
    "langchain-anthropic>=1.4.1",
    "pydantic>=2.13.3",
    "python-dotenv>=1.2.2",
]
